The solution !!
Thanks to `Blake` for the help

```
Blake
5:06 PM (20 minutes ago)

I wonder if it's still possible that your webserver is not started, since you're not using the async start method in your httpserver verticle. The server could not be started before the tests start, because the verticle deploy is completed, but the server has not started yet. Note I still havent run this was just looking at it. I've gotta go right now but will see what's up when i get back in a few hours. In the above code, you'd be getting a null pointer because you were not checking whether or not the AsyncResult was successful or not, so it could have failed and thus the result was null that you tried to call .statusCode() on.
```